userIcon & userIconslice updates
