+++ client/app/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
+import userIconReducer from '../features/userIcon/userIconSlice';
 //https://redux-toolkit.js.org/tutorials/quick-start
 //creating our store - passing in reducers into store
 
 export const store = configureStore({
 	reducer: {
-		counter: counterReducer,
+		userIcon: userIconReducer,
 	},
 });
