+++ client/features/counter/Counter.js
@@ -1,28 +1,28 @@
-import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { decrement, increment } from './counterSlice';
+// import React from 'react';
+// import { useSelector, useDispatch } from 'react-redux';
+// import { decrement, increment } from './counterSlice';
 
-export function Counter() {
-	const count = useSelector((state) => state.counter.value);
-	const dispatch = useDispatch();
+// export function Counter() {
+// 	const count = useSelector((state) => state.counter.value);
+// 	const dispatch = useDispatch();
 
-	return (
-		<div>
-			<div>
-				<button
-					aria-label='Increment value'
-					onClick={() => dispatch(increment())}
-				>
-					Increment
-				</button>
-				<span>{count}</span>
-				<button
-					aria-label='Decrement value'
-					onClick={() => dispatch(decrement())}
-				>
-					Decrement
-				</button>
-			</div>
-		</div>
-	);
-}
+// 	return (
+// 		<div>
+// 			<div>
+// 				<button
+// 					aria-label='Increment value'
+// 					onClick={() => dispatch(increment())}
+// 				>
+// 					Increment
+// 				</button>
+// 				<span>{count}</span>
+// 				<button
+// 					aria-label='Decrement value'
+// 					onClick={() => dispatch(decrement())}
+// 				>
+// 					Decrement
+// 				</button>
+// 			</div>
+// 		</div>
+// 	);
+// }
