+++ client/features/counter/counterSlice.js
@@ -1,32 +1,32 @@
-import { createSlice } from '@reduxjs/toolkit';
+// import { createSlice } from '@reduxjs/toolkit';
 
-const initialState = {
-	value: 0,
-};
+// const initialState = {
+// 	value: 0,
+// };
 
-//https://redux-toolkit.js.org/tutorials/quick-start
+// //https://redux-toolkit.js.org/tutorials/quick-start
 
-export const counterSlice = createSlice({
-	name: 'counter',
-	initialState,
-	reducers: {
-		increment: (state) => {
-			// Redux Toolkit allows us to write "mutating" logic in reducers. It
-			// doesn't actually mutate the state because it uses the Immer library,
-			// which detects changes to a "draft state" and produces a brand new
-			// immutable state based off those changes
-			state.value += 1;
-		},
-		decrement: (state) => {
-			state.value -= 1;
-		},
-		incrementByAmount: (state, action) => {
-			state.value += action.payload;
-		},
-	},
-});
+// export const counterSlice = createSlice({
+// 	name: 'counter',
+// 	initialState,
+// 	reducers: {
+// 		increment: (state) => {
+// 			// Redux Toolkit allows us to write "mutating" logic in reducers. It
+// 			// doesn't actually mutate the state because it uses the Immer library,
+// 			// which detects changes to a "draft state" and produces a brand new
+// 			// immutable state based off those changes
+// 			state.value += 1;
+// 		},
+// 		decrement: (state) => {
+// 			state.value -= 1;
+// 		},
+// 		incrementByAmount: (state, action) => {
+// 			state.value += action.payload;
+// 		},
+// 	},
+// });
 
-// Action creators are generated for each case reducer function
-export const { increment, decrement, incrementByAmount } = counterSlice.actions;
+// // Action creators are generated for each case reducer function
+// export const { increment, decrement, incrementByAmount } = counterSlice.actions;
 
-export default counterSlice.reducer;
+// export default counterSlice.reducer;
