+++ client/features/userIcon/UserIcon.js
@@ -1,26 +1,17 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { decrement, increment } from './userIconSlice';
+import { selectUser } from './userIconSlice';
 
 export function UserIcon() {
-  const count = useSelector((state) => state.usericon.value);
+  const status = useSelector((state) => state.userIcon.addedToSession);
   const dispatch = useDispatch();
 
   return (
     <div>
       <div>
-        <button
-          aria-label='Increment value'
-          onClick={() => dispatch(increment())}
-        >
-          Increment
-        </button>
-        <span>{count}</span>
-        <button
-          aria-label='Decrement value'
-          onClick={() => dispatch(decrement())}
-        >
-          Decrement
+        <button className='userButton' onClick={() => dispatch(selectUser())}>
+          {userName}
+          {status}
         </button>
       </div>
     </div>
