+++ client/features/userIcon/userIconSlice.js
@@ -1,7 +1,8 @@
 import { createSlice } from '@reduxjs/toolkit';
 
 const initialState = {
-  value: 0,
+  addedToSession: new Map(),
+  userName: //get username from wrapper 
 };
 
 //https://redux-toolkit.js.org/tutorials/quick-start
@@ -11,17 +12,22 @@ export const userIconSlice = createSlice({
   name: 'userIcon',
   initialState,
   reducers: {
-    increment: (state) => {
+    selectUser: (state) => {
       // Redux Toolkit allows us to write "mutating" logic in reducers. It
       // doesn't actually mutate the state because it uses the Immer library,
       // which detects changes to a "draft state" and produces a brand new
       // immutable state based off those changes
-      state.value += 1;
+  
+      if (state.addedToSession.has(userName)) {
+        state.addedToSession.delete();
+      } else {
+        state.addedToSession.set(userName, true);
+      }
     },
   },
 });
 
 // Action creators are generated for each case reducer function
-export const { increment } = userIconSlice.actions;
+export const { selectUser } = userIconSlice.actions;
 
 export default userIconSlice.reducer;
