+++ client/features/userIconContainer/UserIconContainer.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { decrement, increment } from './counterSlice';
+
+export function Counter() {
+
+    
+
+
+	const count = useSelector((state) => state.counter.value);
+	const dispatch = useDispatch();
+
+	return (
+		<div>
+			<div>
+				<button
+					aria-label='Increment value'
+					onClick={() => dispatch(increment())}
+				>
+					Increment
+				</button>
+			</div>
+		</div>
+	);
+}
