+++ package.json
@@ -1,7 +1,10 @@
 {
   "dependencies": {
     "@babel/plugin-transform-runtime": "^7.21.0",
+    "@react-google-maps/api": "^2.18.1",
     "@reduxjs/toolkit": "^1.9.3",
+    "bcryptjs": "^2.4.3",
+    "cookie-parser": "^1.4.6",
     "css-loader": "^6.7.3",
     "express": "^4.18.2",
     "load-google-maps-api": "^2.0.2",
